[refactor/#21] Update import paths after renaming auth.ts to authApi.ts
