+++ src/components/layout/UserMenu.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { Link } from "react-router-dom";  
-import { logout } from "../../api/auth.ts"; 
+import { logout } from "../../api/authApi.ts"; 
 import "./UserMenu.css";
 
 export const UserMenu: React.FC = () => {
